% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{cat_session_markdown}
\alias{cat_session_markdown}
\title{Print the Session Information Markdown}
\usage{
cat_session_markdown(f = stdout())
}
\arguments{
\item{f}{File object to append markdown to (stdout by default).}
}
\value{
A string representing the rmarkdown needed for session information.
}
\description{
Session information can be printed in a rmarkdown report. This
function printss the necessary text to make it look nice in the markdown.
}
\details{
Session information is an important part of data science reports, since it
provides the package information and system configuration when the report
was originally generated.

This function prints to the screen the necessary rmarkdown that can be
included in the source of a report. When the report is rendered, the
session information will be included.
}
\note{
This function does not generate a report, only the markdown needed
to generate such a report during \code{\link[rmarkdown]{render}}.
}
\examples{
cat_session_markdown()
}
