% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create package for templates.}
\usage{
create_package(tmpl, from = "dsreportr::ds_pdf")
}
\arguments{
\item{tmpl}{A template of the form \code{pkg::template_pdf}.}

\item{from}{Source package, defaults to \code{dsreportr::ds_pdf}.}
}
\value{
Nothing is returned, but a package should be created.
}
\description{
This function will create a new package (called \code{tmpl}) in
the current directory, then populate the necessary components to be a
R template package.
}
\details{
For RStudio to recognize a package as suitable for a markdown template, there
are several files/directories that are needed. This function will create the
package and then populate the necessary information based on \code{tmpl}. Note
that there remains some work to do with the package, not everything in life is
free! Some things to consider tweaking are the DESCRIPTION (this is not modified)
and the actual template.yaml file used in RStudio.
}
\examples{
\dontrun{create_package("bbireportr::bbi_pdf")}

}
