% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_skeleton_markdown}
\alias{create_skeleton_markdown}
\title{Create skeleton markdown template file}
\usage{
create_skeleton_markdown(tmpl, banner = FALSE)
}
\arguments{
\item{tmpl}{A template of the form \code{pkg::template_pdf}.}

\item{banner}{Should banner be included (default FALSE)}
}
\value{
Nothing, but skeleton created.
}
\description{
This function will copy over the dsreportr template skeleton file and
modify a few variables for it's use in the new template.
}
\note{
This is an internal package function, not for global consumption.
}
