% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_template_in_package}
\alias{create_template_in_package}
\title{Create template files in package in cwd}
\usage{
create_template_in_package(tmpl, from = "dsreportr::ds_pdf")
}
\arguments{
\item{tmpl}{A template of the form \code{pkg::template_pdf}.}

\item{from}{Source package, defaults to \code{dsreportr::ds_pdf}.}
}
\value{
Nothing, but the files will be populated in the existing package.
}
\description{
Populate the necessary template files in the package
that is at the current working directory.
}
\details{
There are a few files/directories needed to make RStudio recognize a package as
templates. This does the work. See \code{\link{create_package}} for details.
}
\examples{
\dontrun{create_template_in_package("bbireportr::bbi_pdf")}
}
