% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_wrapper_function}
\alias{create_wrapper_function}
\title{Create the wrapper pdfdocument function}
\usage{
create_wrapper_function(tmpl, from)
}
\arguments{
\item{tmpl}{A template of the form \code{pkg::template_pdf}.}

\item{from}{A template of the form \code{pkg::template_pdf}.}
}
\value{
Nothing, but the wrapper function is created.
}
\description{
Create the wrapper pdfdocument function
}
