% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_yml_description}
\alias{create_yml_description}
\title{Create a default YML description file}
\usage{
create_yml_description(
  tmpl,
  name = identify_template(tmpl)$template,
  description = glue::glue("A template for {name} based on dsreportr."),
  create_dir = FALSE
)
}
\arguments{
\item{tmpl}{A template of the form \code{pkg::template_pdf}.}

\item{name}{Name of the template (defaults to the tmpl name)}

\item{description}{Description of the template (defaults to A template for name based on dsreportr).}

\item{create_dir}{Should a new markdown file involve a new directory (defaults to FALSE)}
}
\value{
Nothing is return
}
\description{
A YML description file \code{template.yaml} is needed to have RStudio
recognize the package as a template. This function performs that task.
}
\details{
See \code{\link{create_package}} for details.
}
\note{
This is an internal package function, not for global consumption.
}
\examples{
\dontrun{
create_yml_description("bbireportr::bbi_pdf")}
}
