% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{docx}
\alias{docx}
\title{Return docx template as officer object}
\usage{
docx(tmpl = "dsreportr::ds_docx")
}
\arguments{
\item{tmpl}{A string representing a rmarkdown template, with the package
prefix. For example, \code{dsreportr::ds_docx}.}
}
\value{
A docx object for use with officer
}
\description{
Word templates have a template docx file that can be used to start interacting
with the \code{officer} package. This function returns a template object.
}
\examples{
\dontrun{
docx("ds_reportr::ds_docx")
}
}
