% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{docx_template}
\alias{docx_template}
\title{Return path to template docx}
\usage{
docx_template(tmpl = "dsreportr::ds_docx")
}
\arguments{
\item{tmpl}{A string representing a rmarkdown template, with the package
prefix. For example, \code{dsreportr::ds_docx}.}
}
\value{
A path to a word template docx.
}
\description{
Word templates have a template docx file that markdown starts
from for the report. This function returns the path to that word template.
}
\examples{
\dontrun{
docx_template("ds_reportr::ds_docx")
}
}
