% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{encode_library}
\alias{encode_library}
\title{Encode library string as index}
\usage{
encode_library(l)
}
\arguments{
\item{l}{A list of library entries (some of which may be NA)}
}
\value{
A encoded library list (the index of the factor).
}
\description{
The library in sessioninfo is a path, but as a factor. This
translates to numeric index.
}
\details{
The sessioninfo library provides code for translating a library
string into its index so it can be printed succinctly. This duplicates that
mostly unavailable code for use here.
}
