% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_environment_information}
\alias{get_environment_information}
\title{Return R runtime environment}
\usage{
get_environment_information()
}
\value{
The \code{\link[tibble]{tibble}} representing environmental variables.
}
\description{
Provide R runtime environment as a tibble
}
\details{
For markdown templates, it is convenient to print out the environment
during the execution. This includes username, platform, R version,
etc. Most of this exists in \code{\link[utils]{sessionInfo}},
\code{\link[sessioninfo]{platform_info}},
\code{]link[whoami]{username}}. This function simply ties the information
together in a single table for output.
}
\note{
This function returns a table, the corresponding
\code{\link{style_environment_information_as_kable}} or
\code{\link{style_environment_information_as_flextable}}
outputs it for knitting.
}
\examples{
get_environment_information()

}
