% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_package_information}
\alias{get_package_information}
\title{Retrieve package information.}
\usage{
get_package_information()
}
\value{
A tibble representing package information.
}
\description{
Retrieve package information.
}
