% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_template}
\alias{identify_template}
\title{Identify pkg and template from a pkg::template string}
\usage{
identify_template(s)
}
\arguments{
\item{s}{A string of the form pkg::template}
}
\value{
A list of package and template strings, representing the input.
}
\description{
Identify pkg and template from a pkg::template string
}
\examples{
identify_template("dsreportr::ds_pdf")

}
