% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_datestamp.R
\name{knit_with_datestamp}
\alias{knit_with_datestamp}
\title{Knit a document with datestamp appended to output file}
\usage{
knit_with_datestamp(input, output_dir = "delivery", ...)
}
\arguments{
\item{input}{The input file to knit}

\item{output_dir}{An output directory to store results (default is project roo/delivery)}

\item{...}{Other parameters to pass to rmarkdown::render}
}
\value{
rmarkdown::render return result
}
\description{
Using rmarkdown::render to generate an output file from a markdown file.
This function includes a datestamp on the filename to make it unique (by day).
}
\details{
Generating output from markdown document overwrites the existing output file.
Source markdown can be versioned, but it would be convenient to also store multiple copies
of a report. This function appends a date stamp (YYYY-MM-DD) to the output file so
reports can change over time. Ideally, this would follow the versioning in git but that
has not been implemented yet.
}
\note{
Knitr needs the yml header to be simple (or extra syntax is required), so this
function is designed to operate without specifying parameters in the yml.
}
\examples{
\dontrun{
knit_with_datestamp
}
}
