% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_template.R
\name{latex_template}
\alias{latex_template}
\title{Return path to latex template file}
\usage{
latex_template()
}
\value{
A path to a template latex file, suitable for using in rmarkdown
reports.
}
\description{
This returns the path to the latex template file that can
used in this package or other packages that derive from it.
}
\examples{
latex_template()

}
