% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_markdown.R
\name{new_markdown}
\alias{new_markdown}
\title{Create markdown file based on existing template.}
\usage{
new_markdown(tmpl = "dsreportr::ds_pdf", name = "Untitled.Rmd")
}
\arguments{
\item{tmpl}{A string representing a rmarkdown template, with the package
prefix. For example, "dsreportr::ds_pdf".}

\item{name}{A filename to write out the markdown to (default is Untitled.Rmd).}
}
\value{
Logical value indicating success/failure of the file copy operation (
see \code{\link[base]{file.copy}}).
}
\description{
Markdown templates have skeleton markdown files that can
be used to start a report. This function creates a markdown file based
on that template.
}
\examples{
\dontrun{new_markdown("dsreportr::ds_pdf")}

}
