% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{print_session_information}
\alias{print_session_information}
\title{Print R Session information as formatted output.}
\usage{
print_session_information()
}
\value{
Nothing is returned, side effect is printed kable's.
}
\description{
This function will print out the R sessioninfo() information
and some other variables in several tables for reproducibility.
}
\details{
NB: This turns out to be difficult because the individual functions
need to have the kable call as the last thing (so I think an implicit
return) but \emph{cannot} have a return() statement. The calling function
should store the values, then print() the results. Clearly there are
several functions being too clever by half in the stack.
}
