% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{style_environment_information_as_flextable}
\alias{style_environment_information_as_flextable}
\title{Style environment information as flextable output.}
\usage{
style_environment_information_as_flextable(tbl)
}
\arguments{
\item{tbl}{A tibble of environment information (see
\code{\link{get_environment_information}}).}
}
\value{
A \code{\link[flextable]{flextable}} styled for output.
}
\description{
Style environment information as flextable output.
}
\examples{
style_environment_information_as_flextable(get_environment_information())
}
