% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{style_environment_information_as_kable}
\alias{style_environment_information_as_kable}
\title{Style environment information as kable output.}
\usage{
style_environment_information_as_kable(tbl)
}
\arguments{
\item{tbl}{A tibble of environment information (see
\code{\link{get_environment_information}}).}
}
\value{
A \code{\link[knitr]{kable}} styled for output.
}
\description{
This function styles the output of an environment table suitable
for inclusion in a markdown report.
}
\examples{

style_environment_information_as_kable(get_environment_information())

}
