% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{style_package_information_as_flextable}
\alias{style_package_information_as_flextable}
\title{Style package information as flextable.}
\usage{
style_package_information_as_flextable(packages)
}
\arguments{
\item{packages}{The packages tibble from
\code{\link[sessioninfo]{session_info}}}
}
\value{
A \code{\link[flextable]{flextable}} formatted object.
}
\description{
Provide a \code{\link[flextable]{flextable}} formatted output
for the package information.
}
\examples{
style_package_information_as_flextable(get_package_information())
}
