% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{style_package_information_as_kable}
\alias{style_package_information_as_kable}
\title{Style package information as knitr kable.}
\usage{
style_package_information_as_kable(packages)
}
\arguments{
\item{packages}{The packages tibble from
\code{\link[sessioninfo]{session_info}}}
}
\value{
Kable-formatted package information.
}
\description{
Provide a \code{\link[knitr]{kable}} formatted output
for the package information.
}
\examples{
style_package_information_as_kable(get_package_information())

}
