% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_info}
\alias{template_info}
\title{Extract template information from a template.}
\usage{
template_info(f)
}
\arguments{
\item{f}{A path to a \code{template.yaml} file.}
}
\value{
A data frame with template information.
}
\description{
A template is defined in part by a YAML file containing template
details. This function extracts the contents of the YAML and returns in a tibble.
}
\examples{
\dontrun{
template_info("/path/to/pkg/rmarkdown/templates/pkg/template.yaml")

## # A tibble: 1 × 3
##     template name                            description
##     <chr>    <chr>                           <chr>
## 1   ds_docx  Data Science DOCX Report Output A Generic Data Science DOCX report.

}
}
