% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{use_session_markdown}
\alias{use_session_markdown}
\title{Output session markdown text to file}
\usage{
use_session_markdown(f)
}
\arguments{
\item{f}{A filename to append session markdown information to.}
}
\value{
Nothing
}
\description{
In the spirit of the \code{usethis} package, output the
text for including session information into the end of the named
markdown file.
}
