% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_template_in_package}
\alias{create_template_in_package}
\title{Create template files in package in cwd}
\usage{
create_template_in_package(tmpl, banner = NULL, from = "dsreportr::ds_pdf")
}
\arguments{
\item{tmpl}{A template of the form \code{pkg::template_pdf}.}

\item{banner}{A image file to use as banner (see \code{\link{use_banner}}).}

\item{from}{Source package, defaults to \code{dsreportr::ds_pdf}.}
}
\value{
Nothing, but the files will be populated in the existing package.
}
\description{
Populate the necessary template files in the package
that is at the current working directory.
}
\details{
For RStudio to recognize a package as suitable for a
markdown template, there are several files/directories that are needed. This function will create the
package and then populate the necessary information based on \code{tmpl}. Note
that there remains some work to do with the package, not everything in life is
free! Some things to consider tweaking are the DESCRIPTION (this is not modified)
and the actual template.yaml file used in RStudio.
}
\examples{
\dontrun{create_template_in_package("bbireportr::bbi_pdf")}
}
